bug cost volume on cpu solved

+++ prova_dr/src/image.cpp
@@ -2,6 +2,6 @@
 
 int mseBetween2Colors(cv::Vec3b& clr1, cv::Vec3b& clr2){
   // int mse=(pow((int)clr1(0)-(int)clr2(0),2)+pow((int)clr1(1)-(int)clr2(1),2)+pow((int)clr1(2)-(int)clr2(2),2));
-  int mse=(((int)clr1(0)-(int)clr2(0))*((int)clr1(0)-(int)clr2(0))+((int)clr1(1)-(int)clr2(1))*((int)clr1(1)-(int)clr2(1))+((int)clr1(2)-(int)clr2(2))*((int)clr1(2)-(int)clr2(2)));
+  int mse=(clr1(0)-clr2(0))*(clr1(0)-clr2(0))+(clr1(1)-clr2(1))*(clr1(1)-clr2(1))+(clr1(2)-clr2(2))*(clr1(2)-clr2(2));
   return mse;
 }
